refactor(desktop-core): split AutoMobileContent concerns
